requirements.txt
python-telegram-bot>=20.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
easyocr>=1.7.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
